Unstructured methods belong in unstructured.go
Simple code move
